refactor(core): move set attribute code to a DOM-dedicated place
Moves the code responsible for setting an attribute value out of the
instructions folder into the dedicated DOM-manipulation logic file.
