+++ packages/core/src/render3/dom_node_manipulation.ts
@@ -12,6 +12,8 @@ import {escapeCommentText} from '../util/dom';
 import {assertDefined, assertString} from '../util/assert';
 import {setUpAttributes} from './util/attrs_utils';
 import {TNode} from './interfaces/node';
+import {SanitizerFn} from './interfaces/sanitization';
+import {renderStringify} from './util/stringify_utils';
 
 export function createTextNode(renderer: Renderer, value: string): RText {
   return renderer.createText(value);
@@ -93,6 +95,28 @@ export function nativeRemoveNode(renderer: Renderer, rNode: RNode, isHostElement
  */
 export function clearElementContents(rElement: RElement): void {
   rElement.textContent = '';
+}
+
+/**
+ * Set a value of a DOM element attribute. An attribute is removed if the value is null or undefined.
+ */
+export function setElementAttribute(
+  renderer: Renderer,
+  element: RElement,
+  namespace: string | null | undefined,
+  tagName: string | null,
+  name: string,
+  value: any,
+  sanitizer: SanitizerFn | null | undefined,
+) {
+  if (value == null) {
+    renderer.removeAttribute(element, name, namespace);
+  } else {
+    const strValue =
+      sanitizer == null ? renderStringify(value) : sanitizer(value, tagName || '', name);
+
+    renderer.setAttribute(element, name, strValue as string, namespace);
+  }
 }
 
 /**
